+++ changelog.md
@@ -11,6 +11,8 @@ Codion Change Log
 - FilterComboBoxModel.FilterComboBoxModelItems renamed FilterComboBoxItems.
 - FilterComboBoxModel.validSelectionPredicate() moved to FilterComboBoxSelectionModel and renamed validPredicate().
 - FilterComboBoxModel.filterSelectedItem() moved to FilterComboBoxSelectionModel and renamed filterSelected().
+### is.codion.swing.common.ui
+- FilterTableSearchModel.searchPredicate() renamed predicate().
 ### is.codion.framework.model
 - EntityComboBoxModel.linkForeignKeyFilterComboBoxModel() and linkForeignKeyConditionComboBoxModel() replaced with foreignKeyComboBoxModelLinker() returning a ForeignKeyComboBoxModelLinker.
 ### is.codion.swing.framework.model